refactored vulkan and directx 11 code into namespaces
